requirements.txt
aiogram==3.1.1
asyncio~=3.4.3
APScheduler==3.10.4
python-dotenv==1.0.0